hn fixed, #f and #g fixed

+++ srcs/ft_format_functions/ft_format_n.c
@@ -6,7 +6,7 @@ int	ft_format_n(t_printf data, t_flags flag, va_list ap)
 	if (ft_search(data.s, "h") && *(ft_search(data.s, "h") - 1) == 'h')
 		*(va_arg(ap, char *)) = data.ret;
 	else if (ft_search(data.s, "h"))
-		*(va_arg(ap, int *)) = data.ret;
+		*(va_arg(ap, short int *)) = data.ret;
 	else if (ft_search(data.s, "l") && *(ft_search(data.s, "l") - 1) == 'l')
 		*(va_arg(ap, long long int *)) = data.ret;
 	else if (ft_search(data.s, "l"))
